Move user config schema to config package
